dev-cea-consolidation: moved country name query to query helper instead of handler.
