+++ Agents/CEAAgent/cea-agent/src/main/java/uk/ac/cam/cares/jps/agent/cea/utils/geometry/GeometryHandler.java
@@ -20,14 +20,9 @@ import org.apache.jena.geosparql.implementation.parsers.wkt.WKTReader;
 import org.locationtech.jts.operation.buffer.BufferOp;
 import org.locationtech.jts.operation.buffer.BufferParameters;
 
-import org.json.JSONObject;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 
 public class GeometryHandler {
     public static final String EPSG_4326 = "EPSG:4326";
@@ -279,40 +274,4 @@ public class GeometryHandler {
         return crsList.stream().allMatch(s -> s.equals(crsList.get(0)));
     }
 
-    public static String getCountry(Coordinate coordinate) {
-        String urlString = String.format(
-                "https://nominatim.openstreetmap.org/reverse?lat=%f&lon=%f&format=json",
-                coordinate.getY(), coordinate.getX());
-
-        try {
-            URL url = new URL(urlString);
-            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-            conn.setRequestMethod("GET");
-
-            conn.setRequestProperty("User-Agent", "CEA agent");
-
-            // Read the response
-            BufferedReader reader = new BufferedReader(new InputStreamReader(conn.getInputStream()));
-            StringBuilder response = new StringBuilder();
-            String line;
-
-            while ((line = reader.readLine()) != null) {
-                response.append(line);
-            }
-            reader.close();
-
-            JSONObject j = new JSONObject(response.toString());
-
-            if (j.has("address")) {
-                if (j.getJSONObject("address").has("country_code")) {
-                    return j.getJSONObject("address").getString("country_code");
-                }
-            }
-
-            return "CH";
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "CH";
-        }
-    }
 }
